1st question completed - needs more tuning

+++ assignment-2-filtering/1/code/get_kernel_radius.m
@@ -14,7 +14,7 @@ function kr_img = get_kernel_radius(mask, thresh_dist)
 %
 %%
 %
-kr_img = bwdist(mask);
+kr_img = double(bwdist(mask));
 kr_img(kr_img > thresh_dist) = thresh_dist;
 figure;
 sgtitle('Part (c) - (ii) Contour Plot');
